test: add to post processing test and move to misc file
